% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPlot.R
\name{ColorPlot}
\alias{ColorPlot}
\title{ColorPlot display multiple RGB colors}
\usage{
ColorPlot(colors.vec, n.col, n.row, label = NULL)
}
\arguments{
\item{n.col}{number of columns for the color matrix}

\item{colors.ved}{a vector with RGB colors in format "#354C31"}
}
\description{
To make picking RGB color set easier, this function display multiple colors in a matrix. Custom labels
}
