% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot.grey.R
\name{DimPlot.grey}
\alias{DimPlot.grey}
\title{DimPlot with grey background in split panels}
\usage{
DimPlot.grey(
  object,
  dims = c(1, 2),
  pt.size = 0.2,
  reduction = "umap",
  group.by = NULL,
  split.by = NULL,
  shape.by = NULL,
  shuffle = F,
  ncol = NULL,
  back.col = "lightgrey"
)
}
\arguments{
\item{...}{same as Seurat DimPlot function.}
}
\description{
In Seurat DimPlot function, split.by parameter allows cells to be splited into 
separate panels according to a meta column. However, it's sometimes hard to compare across panels because
some clusters are not existed in some conditions, and therefore the look of embeddings in each panel can 
look drastically different.

To locate cells on the full embedding easier, this function plot cells in other panels in grey at the back,
providing a shape of the full embedding.
}
\examples{
DimPlot.grey(srt, group.by = "Disease", split.by = "Disease")
}
