% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot.patient.R
\name{DimPlot.patient}
\alias{DimPlot.patient}
\title{DimPlot.patient}
\usage{
DimPlot.patient(
  object,
  dims = c(1, 2),
  pt.size = 0.2,
  reduction = "umap",
  group.by = NULL,
  split.by = NULL,
  shape.by = NULL,
  shuffle = F,
  ncol = NULL,
  grey = T
)
}
\arguments{
\item{grey}{Add cells belong to other panels as grey at the back, if TRUE.}

\item{...}{same as Seurat DimPlot function.}
}
\description{
DimPlot with grey background in split panels, and color each panel independently.
 When there are too many catagories in group.by, colors can be too similar and therefore hard
 to see their location on embedding plot. When we can put them into some meaningful groups, e.g.
 Group patients by disease, we can split.by group and use a whole color scale for each panel.
}
\examples{
DimPlot.patient(srt, group.by = "Patient", split.by = "Disease")
}
