% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot.ssc.R
\name{DotPlot.ssc}
\alias{DotPlot.ssc}
\title{DotPlot.ssc with sample mean as each dot.}
\usage{
DotPlot.ssc(
  input,
  gene,
  group.by,
  split.by = NULL,
  point.by = "Patient",
  num.cutoff = 10,
  title = "",
  number_labels = T,
  text_sizes = c(20, 10, 5, 10, 5, 5, 3),
  spread = NULL,
  theme = "classic",
  jitter.alpha = 0.5,
  jitter.width = 0.2,
  jitter.size = 2,
  col.trans = "",
  bar = T,
  link = FALSE
)
}
\arguments{
\item{input}{Seurat object.}

\item{group.by}{Group and color by this meta column}

\item{split.by}{Split into different panels. A vector with one or two elements, in order of c(row, col)}

\item{point.by}{"Patient" by default.}

\item{num.cutoff}{Color dots with lower than num.cutoff cells to grey.}

\item{title}{Plot title.}

\item{number_labels}{Show stats at the bottom of the plot if TRUE. First row is total cell number in each condition,
second row is fraction of cells with at least one UMI of this gene detected.}

\item{text_sizes}{Seven elements in order: plot title, axis title, axis text, legene title, legend text, split, stat size.}

\item{spread}{Control skin is only in HealthyControl in Disease meta column, but we want to compare them to other skins in all diseases
even when we split by Disease. So we spread HealthyControl condition in Disease column, by spread = c("Disease", "HealthyControl").}

\item{theme}{ggplot theme. classic by default.}

\item{col.trans}{Set to "log" to make color scale zoom in to lower values}

\item{bar}{Plot weighted mean as a bar if TRUE.}

\item{link}{Link the paired conditions in each panel. However, the starting and ending points cannot adjust according to
the horizontal location of jitter plot. Can set a lower value for jitter.width to shrink the blank distance.}
}
\description{
To evaluate a gene's consistency across patients, it's useful to know the expression
level of individual samples. If set point.by to Patient, and use group.by and split.by to separate conditions, each
dot in this plot would represent the average mean of each patient. Color by the number of cells used for mean calculation,
indicating how trustable they are. Blue at the lower end, then red to orange. Set bar to TRUE to plot the condition mean
with all cells from all samples included.
}
\details{
Use mean of normalized UMI of cells in a condition instead of aggregated bulk to simplify calculation.
}
\examples{
srt$Skin <- factor(srt$Skin, levels = c("Control", "NonLesional", "Lesional"))
DotPlot.ssc(srt, "FOSL1", group.by = "Skin", split.by = c("subCellType","DiseaseFiner"), spread = c("DiseaseFiner","HealthyControl"))
}
