% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapPlot.ssc.R
\name{HeatmapPlot.ssc}
\alias{HeatmapPlot.ssc}
\title{Heatmap with single cell level or mean expression value.}
\usage{
HeatmapPlot.ssc(
  object,
  genes,
  group.by = NULL,
  split.by = NULL,
  color = NULL,
  assay = "RNA",
  slot = "data",
  dataScale = "row",
  title = NULL,
  showRowName = T,
  showColName = T,
  clusterRow = FALSE,
  clusterCol = F,
  reorderRow = F,
  reorderCol = F,
  annRightDf = NULL,
  annTopShow = F,
  row_split_str = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{genes}{A vector of genes for heatmap.}

\item{group.by}{Metadata column to aggregate for mean expression value. Null for single-cell.}

\item{split.by}{Split into columns.}

\item{color}{User defined color panel for ComplexHeatmap.}

\item{assay}{Seurat assay to get gene expression values.}

\item{slot}{Seurat slot of assy for gene expression values.}

\item{title}{Plot title.}
}
\description{
Plot a heatmap with single cell level or mean expression value to show the expression level of genes in the list.
}
\examples{
HeatmapPlot.ssc(srt.ic, genes = c("CLEC9A","XCR1","IFI35","CLEC10A"), group.by = "DiseaseFiner")
}
