% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentPlot.R
\name{PercentPlot.single}
\alias{PercentPlot.single}
\title{percent plot with denominator and devider specified.}
\usage{
PercentPlot.single(
  object,
  numerator,
  numerator.column,
  denominator,
  denominator.column = NULL,
  group.by,
  split.by = NULL,
  point.by = "Patient",
  spread = NULL,
  num.cutoff = 20,
  num.cutoff.rm = F,
  number.labels = T,
  xaxis.labels = F,
  title = "",
  theme = "classic",
  text.size.plot.title = 20,
  text.size.axis.title = 10,
  text.size.axis.text = 10,
  text.size.legend.title = 10,
  text.size.legend.text = 5,
  text.size.panel = 5,
  jitter.alpha = 0.5,
  jitter.width = 0.1,
  jitter.size = 2,
  jitter.col = T,
  col.trans = "identity",
  bar = T,
  link = FALSE,
  ttest = F
)
}
\arguments{
\item{object}{Seurat object.}

\item{numerator}{One type to calculate percentage for, e.g. CD4}

\item{numerator.column}{Name of the metadata column where the numerator string is stored.}

\item{denominator}{A vector with all types to be used as total cells for percentage calculation.}

\item{denominator.column}{Name of the metadata column where the denominator strings are stored. Same as numerator.column if not defined.}

\item{group.by}{Group and color by this meta column}

\item{split.by}{Split into different panels. A vector with one or two elements, in order of c(row, col)}

\item{point.by}{"Patient" by default.}

\item{spread}{Control skin is only in HealthyControl in Disease meta column, but we want to compare them to other skins in all diseases
even when we split by Disease. So we spread HealthyControl condition in Disease column, by spread = c("Disease", "HealthyControl").}

\item{num.cutoff}{Color dots with fewer cells in denominator than num.cutoff in grey.}

\item{num.cutoff.rm}{Keep the grey dots not passing the num.cutoff threshold or remove them.}

\item{title}{Plot title.}

\item{theme}{ggplot theme. classic by default.}

\item{jitter.alpha}{Alpha of dots.}

\item{jitter.width}{Spread out width of dots.}

\item{jitter.size}{Size of dots.}

\item{jitter.col}{Whether to color dots based on number of cells within, or leave dots black.}

\item{col.trans}{Set to "log" to make color scale zoom in to lower values}

\item{bar}{Plot weighted mean as a bar if TRUE.}

\item{link}{Link the paired conditions in each panel. However, the starting and ending points cannot adjust according to
the horizontal location of jitter plot. Can set a lower value for jitter.width to shrink the blank distance.}

\item{number_labels}{Show stats at the bottom of the plot if TRUE. Total cell number in that condition.}

\item{text_sizes}{Seven elements in order: plot title, axis title, axis text, legene title, legend text, split, stat size.}
}
\description{

}
\examples{
PercentPlot.single(srt, numerator="CD8_RDH10", numerator.column="subCellType.ificc", denominator="CD8", denominator.column="subCellType.ificc.main", group.by = "Skin", split.by = "DiseaseFiner", ttest=F, link = T, num.cutoff.rm=T)
}
