% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiePlot.ssc.R
\name{PiePlot.ssc}
\alias{PiePlot.ssc}
\title{percent plot with denominator and divider specified.}
\usage{
PiePlot.ssc(
  object,
  group.by,
  split.by = NULL,
  layer.by = NULL,
  normalize.by = NULL,
  ind.cell = NULL,
  width = 0.9,
  title = NULL,
  number_labels = "percentage",
  label.digit = 4,
  plot.theme = "void",
  color = NULL
)
}
\arguments{
\item{object}{Seurat object. Full dataset for normalization denominators.}

\item{group.by}{Color by this meta column.}

\item{split.by}{Split into different panels.}

\item{layer.by}{Multiple layers of rings in each pie chart.}

\item{normalize.by}{A vector with metadata column names for normalizing groups. Use proportion as pie chart value. For example, some diseases have more cells in total, normalized by disease and pie chart for prevalence in each disease. NULL to plot the raw value.}

\item{ind.cell}{Indices for cells to show in plot. NULL by default, plotting all conditions.}

\item{width}{Width of each ring. 1 for no gap between rings.}

\item{title}{Plot title.}

\item{number_labels}{Show raw value or percentage in each condition. FALSE, "value", "percentage"}

\item{label.digit}{Number of decimal places to keep.}

\item{plot.theme}{ggplot theme. void by default. NULL to skip.}
}
\description{
This function uses a pie plot to show the percentage of each group in the total. The denominator and divider can be specified to normalize the percentage.
}
\examples{
diseases <- c("HealthyControl", "Dermatomyositis", "Lupus", "Psoriasis", "Vitiligo-active")
skins <- c("Control", "Lesional")
celltypes <- c("MEL","KC_act")
celltypeColName <- "subCellType.ificc"

object <- srt[,srt$Skin \%in\% skins &
                srt$DiseaseFiner \%in\% diseases]

ind <- object@meta.data[,celltypeColName] \%in\% celltypes
PiePlot.ssc(srt, group.by="DiseaseFiner", split.by="subCellType.ificc", normalize.by=c("DiseaseFiner"), ind.cell=ind,
width=0.9, title=NULL, label=T, color=NULL)
}
