% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToMeta.R
\name{ToMeta}
\alias{ToMeta}
\title{ToMeta}
\usage{
ToMeta(genes, data = NULL, to = "mergedname")
}
\arguments{
\item{genes}{A vector of genes.}

\item{data}{The mata gene name table. If NULL, use the default table for FourDisease data.}

\item{to}{Change to meta or mergedgene name. Or use the name of a custom column.}
}
\description{
A wrapper to change a list of genes to their meta gene names if applicable. 
 Return a genelist with new meta gene names and print the part of matagene table describing 
 meta genes included in gene list. Can be directly used for ploting with new data
 print info for meta genes included.
}
\examples{
FeaturePlot(srt, ToMeta(c("CD4","FOXP3")))
}
