% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VlnPlot.compact.R
\name{VlnPlot.compact}
\alias{VlnPlot.compact}
\title{compact violin plots to fit more information in less space}
\usage{
VlnPlot.compact(
  object,
  genes,
  group.by,
  title = "",
  assay = NULL,
  slot = "data",
  log_scale = F,
  colors = NULL,
  split.by = NULL,
  spread = NULL,
  split.dir = "h",
  split.scale = "fixed",
  split.label.pos = "top",
  split.label.textonly = F,
  split.label.hide = "none",
  split.label.rotate = "none",
  axis.hide = "none",
  legend.hide = F,
  flip = F,
  jitter_pts = F,
  plot_mean = F,
  size = 1,
  sig = 3,
  number_labels = F,
  text_sizes = c(15, 10, 7, 10, 7, 7, 2.5),
  alpha = 0.5,
  theme = "classic"
)
}
\arguments{
\item{object}{Seurat object.}

\item{genes}{a vector of genes to plot.}

\item{group.by}{Group and color by this meta column.}

\item{title}{Plot title.}

\item{assay}{Assay of the seurat object for the expression values.}

\item{slot}{Slot of the seurat object for the expression values.}

\item{log_scale}{log2 scale the values or not. By default FALSE.}

\item{split.by}{Split into different panels. A vector with one or two elements, in order of c(row, col)}

\item{spread}{Control skin is only in HealthyControl in Disease meta column, but we want to compare them to other skins in all diseases
even when we split by Disease. So we spread HealthyControl condition in Disease column, by spread = c("Disease", "HealthyControl").}

\item{split.dir}{The direction arranging split panels by one variable. "h" horizontally (default), "v" vertically.}

\item{split.scale}{Split panels share the same axis scale "fixed" (default), or use different scales "free".}

\item{split.label.pos}{When split by two variables, put the universal label at the top and right as default, or "bottom", or "left".}

\item{split.label.textonly}{Remove the rectangular frame from the panel labels. FALSE by default.}

\item{split.label.hide}{Remove panel labels. "none" by default. "x" "y" "both" as options.}

\item{split.label.rotate}{Rotate the label for 90 degrees. "none" by default. "x" "y" "both" as options.}

\item{axis.hide}{Hide axis ticks and text for a simpler look. "none" by default. "x" "y" "both" as options.}

\item{legend.hide}{Hide legend. FALSE by default.}

\item{flip}{Change the violin plots to horizontal direction.}

\item{jitter_pts}{dot for value of each individual cell. FALSE by default.}

\item{plot_mean}{plot the mean expression level as a black dot. FALSE by default.}

\item{sig}{Significant number of the cell proportion with positive values.}

\item{number_labels}{Show stats at the bottom of the plot if TRUE. Total cell number in that condition.}

\item{text_sizes}{Six elements in order: plot title, axis title, axis text, legene title, legend text, split, stat size}

\item{plot.theme}{ggplot theme. classic by default.}
}
\description{
Comparing to the regular VlnPlot.ssc function, a few elements are ignored by default, including cell count in condition, proportion with positive values, dots for individual cell value, and mean expression level.
Supports a few more options including horizontal violin, multiple genes,
}
\examples{
genes <- c("IRF8","CLEC9A","XCR1","CLEC10A","CD14","CD209","CXCL9","IL1B","CCL17","JCHAIN","CD207","CCR7")
VlnPlot.compact(srt, assay = "RNA", genes = genes, group.by = "subCellType.newmel", split.by = c("gene","subCellType.newmel"),
split.scale = "free", split.label.pos = c("left", "bottom"), split.label.textonly = T,
split.label.rotate = "both", axis.hide = "both", flip = T, legend.hide = T)

VlnPlot.compact(srt, genes = "CXCL9", group.by = "subCellType.newmel", legend.hide = T) +
  theme(axis.text.x = element_text(angle = 90))
}
