% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VlnPlot.ssc.R
\name{VlnPlot.ssc}
\alias{VlnPlot.ssc}
\title{VlnPlot.ssc Alternate violin plot function for Seurat object}
\usage{
VlnPlot.ssc(
  object,
  gene,
  group.by,
  group.order = NULL,
  split.by = NULL,
  split.order = NULL,
  title = "",
  assay = "RNA",
  slot = "data",
  log_scale = F,
  colors = NULL,
  spread = NULL,
  jitter_pts = T,
  violin.linewidth = 1,
  plot_mean = T,
  size = 1,
  sig = 3,
  number_labels = T,
  text_sizes = c(15, 10, 7, 10, 7, 7, 2.5),
  alpha = 0.5,
  theme = "classic"
)
}
\arguments{
\item{object}{Seurat object to plot.}

\item{gene}{Gene name for plot. Wrap in ToMeta function if needed.}

\item{group.by}{Group and color cells.}

\item{split.by}{Split into panels. A vector with one or two elements. Take in order as c(row, column).}

\item{title}{Plot title.}

\item{assay}{Assay to extract gene UMI counts.}

\item{slot}{Slot to extract gene UMI counts.}

\item{log_scale}{Put y-axis into log scale if TRUE.}

\item{colors}{A vector of colors for scale_fill_manual / scale_color_manual to generate the gradient scale.}

\item{spread}{Control skin is only in HealthyControl in Disease meta column, but we want to compare them to other skins in all diseases
even when we split by Disease. So we spread HealthyControl condition in Disease column, by spread = c("Disease", "HealthyControl").}

\item{jitter_pts}{Overlay a jitter plot over violin plot if TRUE.}

\item{violin.linewidth}{Width of the violin borders.}

\item{plot_mean}{Plot the mean UMI value in black dot if TRUE.}

\item{size}{Dot size of jitter plot.}

\item{sig}{Number of decimal places in fraction stats.}

\item{number_labels}{Show stats at the bottom of the plot if TRUE. First row is total cell number in each condition,
second row is fraction of cells with at least one UMI of this gene detected.}

\item{text_sizes}{Seven elements in order: plot title, axis title, axis text, legene title, legend text, split, stat size.}

\item{alpha}{Transparency of violin plot.}

\item{theme}{ggplot theme. classic by default.}
}
\description{
Left y-axis is for UMI level for individual cells, right y-axis is for the mean level showed in large black dots.
}
\examples{
VlnPlot.ssc(srt, gene = "CXCL10", group.by = "Skin", split.by = c("CellType", "Disease"), spread = c("Disease", "HealthyControl"), assay = "RNA")
}
