% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VlnPlot.xlabel.R
\name{VlnPlot.xlabel}
\alias{VlnPlot.xlabel}
\title{VlnPlot.scc with group label below each condition.}
\usage{
VlnPlot.xlabel(
  object,
  gene,
  group.by,
  group.order = NULL,
  split.order = NULL,
  title = "",
  assay = "RNA",
  slot = "data",
  log_scale = F,
  colors = NULL,
  split.by = NULL,
  spread = NULL,
  jitter_pts = T,
  plot_mean = T,
  size = 1,
  sig = 3,
  number_labels = T,
  text_sizes = c(15, 10, 7, 10, 7, 7, 2.5),
  alpha = 0.5,
  theme = "classic"
)
}
\arguments{
\item{...}{same as VlnPlot.ssc}
}
\description{
With more possible values in group.by column, colors of groups can become hard to tell apart.
This function is a variant of VlnPlot, using VlnPlot.scc format, with group labels at the bottom.
}
\examples{
VlnPlot.xlabel(srt, "IFNG", group.by = "Patient", split.by = "Disease")
}
